Update source and fix rpmlint warnings
With the new sources, the patch is no longer needed as the rpaths in
MaxScale have been fixed.

The act of installing a logrotate config without implying a dependency
on logrotate caused rpmlint to emit a warning. Additionally, the file
name was not the name of the package or the program which caused another
warning.

The use of SSL_CTX_set_cipher_list() caused a warning but since this is
only called when the user defines a custom cipher list that overrides
the system one, the warning can be safely ignored.

The maxscale_generate_support_info.py would require Python 3 to function
correctly and given the fact that maxctrl is capable of generating
reports at runtime via `maxctrl create report`, the need for this script
is not that high. Getting rid of the script fixes a few warnings as
well.