Update sources to MaxScale 21.06.17
Updated sources to the officially released source code of MaxScale
21.06.17.

The removal of dbfwfilter solves a few rpmlint warnings and given that
it was removed in the 22.08 release of MaxScale, it seems somewhat
pointless to build it.

The fix for the removal of the prelink config file did not get into
21.06.17 so it still needs to be removed. Similarly, the
/usr/bin/maxctrl file has the wrong permissions and needs to be fixed
(MXS-5264).

The executables now all have manuals for them in 21.06.17 and they need
to be installed.